move config files to src/config; add redirectTarget and redirect there if referrer is empty
