+++ jest.config.js
@@ -13,7 +13,7 @@ module.exports = {
   },
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/src/$1',
-    'config': '<rootDir>/config.local.ts',
+    'config': '<rootDir>/src/config/config.local.ts',
     '@nimiq/iqons/dist/iqons-name.min.js': '@nimiq/iqons/dist/iqons-name.cjs.js',
   },
   snapshotSerializers: [
