+++ src/config/config.local.ts
@@ -1,8 +1,9 @@
-import { NETWORK_TEST } from './src/lib/Constants';
+import { NETWORK_TEST } from '../lib/Constants';
 
 export default {
     keyguardEndpoint: window.location.protocol + '//' + window.location.hostname + ':8000/src',
     network: NETWORK_TEST,
     networkEndpoint: 'https://network.nimiq-testnet.com',
     privilegedOrigins: [ '*' ],
+    redirectTarget: window.location.protocol + '//' + window.location.hostname + ':8080/demos',
 };
