+++ src/config/config.mainnet.ts
@@ -1,8 +1,9 @@
-import { NETWORK_MAIN } from './src/lib/Constants';
+import { NETWORK_MAIN } from '../lib/Constants';
 
 export default {
     keyguardEndpoint: 'https://keyguard.nimiq.com',
     network: NETWORK_MAIN,
     networkEndpoint: 'https://network.nimiq.com',
     privilegedOrigins: [ 'https://safe.nimiq.com' ],
+    redirectTarget: 'https://safe.nimiq.com',
 };
