+++ src/config/config.testnet.ts
@@ -1,8 +1,9 @@
-import { NETWORK_TEST } from './src/lib/Constants';
+import { NETWORK_TEST } from '../lib/Constants';
 
 export default {
     keyguardEndpoint: 'https://keyguard.nimiq-testnet.com',
     network: NETWORK_TEST,
     networkEndpoint: 'https://network.nimiq-testnet.com',
     privilegedOrigins: [ 'https://safe.nimiq-testnet.com' ],
+    redirectTarget: 'https://safe.nimiq-testnet.com',
 };
