+++ tsconfig.json
@@ -19,7 +19,7 @@
         "src/*"
       ],
       "config": [
-        "config.local.ts"
+        "src/config/config.local.ts"
       ]
     },
     "lib": [
