+++ tsconfig.local.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "paths": {
       "config": [
-        "config.local.ts"
+        "src/config/config.local.ts"
       ]
     }
   }
