+++ tsconfig.testnet.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "paths": {
       "config": [
-        "config.testnet.ts"
+        "src/config/config.testnet.ts"
       ]
     }
   }
